+++ src/slskd/Application/Core/API/DTO/LoginRequest.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="ServerState.cs" company="slskd Team">
+﻿// <copyright file="LoginRequest.cs" company="slskd Team">
 //     Copyright (c) slskd Team. All rights reserved.
 //
 //     This program is free software: you can redistribute it and/or modify
@@ -15,19 +15,11 @@
 //     along with this program.  If not, see https://www.gnu.org/licenses/.
 // </copyright>
 
-namespace slskd.Management
+namespace slskd.Core.API
 {
-    using System.Net;
-    using Soulseek;
-
-    public class ServerState
+    public class LoginRequest
     {
-        public string Address { get; set; }
-        public IPEndPoint IPEndPoint { get; set; }
-        public SoulseekClientStates State { get; set; }
         public string Username { get; set; }
-        public bool IsConnected => State.HasFlag(SoulseekClientStates.Connected);
-        public bool IsLoggedIn => State.HasFlag(SoulseekClientStates.LoggedIn);
-        public bool IsTransitioning => State.HasFlag(SoulseekClientStates.Connecting) || State.HasFlag(SoulseekClientStates.Disconnecting) || State.HasFlag(SoulseekClientStates.LoggingIn);
+        public string Password { get; set; }
     }
 }
