+++ src/slskd/Common/UserEndPointCache.cs
