+++ src/slskd/Exceptions/ApplicationShutdownException.cs
