+++ src/slskd/Exceptions/IntentionalDisconnectException.cs
