+++ src/slskd/Exceptions/RetryException.cs
