+++ src/slskd/Exceptions/ShareAliasException.cs
