+++ src/slskd/Exceptions/ShareScanInProgressException.cs
