+++ src/slskd/Exceptions/SlskdException.cs
